8 - обработка ошибки при загрузке страницы списка предложений
